+++ btclib/blocks.py
@@ -21,7 +21,7 @@ from btclib import var_bytes, var_int
 from btclib.alias import BinaryData
 from btclib.exceptions import BTClibValueError
 from btclib.hashes import hash256, merkle_root
-from btclib.tx.block_header import BlockHeader
+from btclib.block_header import BlockHeader
 from btclib.tx.tx import Tx
 from btclib.utils import bytesio_from_binarydata, decode_num
